extract searchProductType
