+++ src/PrestaShopBundle/Form/Admin/Sell/Product/SearchedProductItemType.php
@@ -35,7 +35,7 @@ use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class SearchedProductType extends CommonAbstractType
+class SearchedProductItemType extends CommonAbstractType
 {
     /**
      * {@inheritDoc}
